+++ CHANGES.rst
@@ -3,6 +3,7 @@
 
 - Changed from fluid layout to fixed 980px layout for widescreen responsive level.
 - Added timestamp to all assets path for cache expiration.
+- Added thumbnail preview support for Imgur links.
 
 0.2.0
 -----